fix: Always force end transition after stop
We used to rely on a `longStopTimeout` to determine if a stop transition
should be added after the last point. This timeout was shorter than the
(~8min) than the waiting time before a stationary event is triggered
(11min) and was supposed to prevent cases where the user hasn't stopped
long enough to consider adding a stop transition.
However, we noticed with some tests on Android the 11min delay is
actually shortened to ~6min, thus less than the longStopTimeout, causing
the last point never having a stop transition, resulting in the last
trip never uploaded (it will be uploaded at the next upload).

This behaviour was made to prevent unexpected trip splitting, but we
prefer to remove this for now.