fix: Upload motion activity only when mode has changed
We used to upload a motion activity document for each uploaded location
point.
This is problematic as it overloads the server database for nothing and,
more importantly, may have side-effects on the resulting trip with too
many contiguous sections with the same trip.

We now only upload a motion activity when it is the first point or when
the new point has a different mode than before, which is the behaviour
of the openpath client.