refactor: Remove uncessary check
It is not necessary to check if new id is correctly saved, because if it is not, storeId would throw an error.

Also, it is better to call just getId and not getOrCreateId before storing the id, to avoid side effect (like creating an id before updating it which is useless).