feat: Use speed to force transitions
When a large temporal delta is detected bewteen 2 locations points, this
might be symptomatic of a stop event that had been missed.
To cope with this, we used to compute the spatial distance between those
points and force a stop/start transition iff the distance was small
enough ; the reasoning was that the device stayed on the same place long
enough to consider the trip should be stopped. And a long
spatio-temporal distance between points could be typical of a plane
trip.

We noticed some drawbacks with this spatial distance: the threshold is
not easy to set, notably because the iOS geofence, and we had some
issues when the user was just outside the threshold, even when the
temporal distance was very large. And we do not want to set a too large
spatial threshold because this could produce unexpected transitions
after the tracking stoppped for a while (e.g. because of a failing
network in a subway).

So, to solve this, we now force a transition when:
- The temporal distance is really high (12h):
  this seems reasonable even for a long distance flight, as we recently
  noticed that forcing a transition does not always imply a trip cut,
  probably because the server is smart enough to detect a flight
  behaviour.
- The temporal distance is high (30min) and the speed is low (0.1 m/s):
  this should tackle the cases where we missed the stationary event, and
  the distance / duration is too low to be considered as a trip
  continuity.